\name{danes}
\alias{danes}
\docType{data}
\title{Y-STR haplotypes}
\description{185 Y-STR 10 loci haplotypes}
\usage{data(danes)}
\format{
  A data frame with 185 observations on the following 10 loci (\code{n} is the number of times each haplotype has been observed)
  \describe{
    \item{\code{DYS19}}{}
    \item{\code{DYS389I}}{}
    \item{\code{DYS389II}}{}
    \item{\code{DYS390}}{}
    \item{\code{DYS391}}{}
    \item{\code{DYS392}}{}
    \item{\code{DYS393}}{}
    \item{\code{DYS437}}{}
    \item{\code{DYS438}}{}
    \item{\code{DYS439}}{}
    \item{\code{n}}{}
  }
}
\source{
''Y-chromosome STR haplotypes Danes'' by Hallenberg et al (2005), http://www.sciencedirect.com/science/article/pii/S0379073805000095
}
\keyword{datasets}
