% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclapmix-package.R
\docType{package}
\name{disclapmix-package}
\alias{disclapmix-package}
\title{disclapmix}
\description{
Discrete Laplace Mixture Inference using the EM Algorithm. 
A central function is \code{\link{disclapmix_adaptive}} (and the underlying \code{\link{disclapmixfit}}).
}
\author{
Mikkel Meyer Andersen <mikl@math.aau.dk> and 
Poul Svante Eriksen <svante@math.aau.dk>
}
