\name{extractFullBIC}
\alias{extractFullBIC}
\title{Extracts BIC based on full likelihood}
\description{Extracts BIC based on full likelihood from a \code{\link{disclapmixfit}} object}
\usage{extractFullBIC(fit)}
\arguments{
  \item{fit}{\code{\link{disclapmixfit}} object}
}
\value{The BIC value}
\author{Mikkel Meyer Andersen <mikl@math.aau.dk> and Poul Svante Eriksen}

\seealso{
\code{\link{disclapmix}}
\code{\link{disclapmixfit}}
\code{\link{plot.disclapmixfit}}
\code{\link{predict.disclapmixfit}}
\code{\link{extractMarginalBIC}}
}

\keyword{BIC}

