\name{extractMarginalBIC}
\alias{extractMarginalBIC}
\title{Extracts BIC based on marginal likelihood}
\description{Extracts BIC based on marginal likelihood from a \code{\link{disclapmixfit}} object}
\usage{extractMarginalBIC(fit)}
\arguments{
  \item{fit}{\code{\link{disclapmixfit}} object}
}
\value{The BIC value}
\author{Mikkel Meyer Andersen <mikl@math.aau.dk> and Poul Svante Eriksen}

\seealso{
\code{\link{disclapmix}}
\code{\link{disclapmixfit}}
\code{\link{plot.disclapmixfit}}
\code{\link{predict.disclapmixfit}}
\code{\link{extractFullBIC}}
}

\keyword{BIC}

