% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{get_rank}
\alias{get_rank}
\title{Get rank of pair}
\usage{
get_rank(x, haplotype)
}
\arguments{
\item{x}{A \code{ranked_contrib_pairs} object.}

\item{haplotype}{A haplotype.}
}
\description{
Get rank of pair
}
