% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_functions.R
\name{happrobsum_within}
\alias{happrobsum_within}
\title{Haplotypes probability sum}
\usage{
happrobsum_within(object, alleles, ...)
}
\arguments{
\item{object}{a \code{\link{disclapmixfit}} object}

\item{alleles}{matrix defining the allele ranges: a column per locus where
first row is minimum allele and second row is maximum allele}

\item{...}{not used}
}
\description{
Calculate sum of haplotype probabilities where haplotypes are defined by
allele ranges.
}
\seealso{
\code{\link{disclapmix}} \code{\link{disclapmixfit}}
\code{\link{print.disclapmixfit}} \code{\link{summary.disclapmixfit}}
\code{\link{simulate.disclapmixfit}} \code{\link{plot.disclapmixfit}}
\code{\link{predict.disclapmixfit}} %\code{\link{haplotype_diversity}}
\code{\link{clusterdist}} \code{\link{clusterprob}}
}
\keyword{happrobsum}
