% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{plot.ranked_contrib_pairs}
\alias{plot.ranked_contrib_pairs}
\title{Plot ranked contributor pairs}
\usage{
\method{plot}{ranked_contrib_pairs}(x, top = NULL, ..., xlab = "Rank",
  ylab = "P(H1)P(H2)")
}
\arguments{
\item{x}{A \code{ranked_contrib_pairs} object.}

\item{top}{The top ranked number of pairs to print. \code{NULL} for
all.}

\item{\dots}{Delegated to the generic \code{\link[graphics]{plot}} function.}

\item{xlab}{Graphical parameter.}

\item{ylab}{Graphical parameter.}
}
\description{
Plot ranked contributor pairs
}
