% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{print.contrib_pairs}
\alias{print.contrib_pairs}
\title{Print contributor pairs}
\usage{
\method{print}{contrib_pairs}(x, ...)
}
\arguments{
\item{x}{A \code{contrib_pairs} object.}

\item{\dots}{Ignored}
}
\description{
Print contributor pairs
}
