\name{print.disclapmixfit}
\alias{print.disclapmixfit}
\title{Print a disclapmixfit}
\description{Prints a \code{\link{disclapmixfit}} object.}
\usage{
  \method{print}{disclapmixfit}(x, \dots)
}
\arguments{
  \item{x}{a \code{\link{disclapmixfit}} object, usually from a result of a call to \code{disclapmix}.}
  \item{...}{not used}
}
\author{
  Mikkel Meyer Andersen <mikl@math.aau.dk> and Poul Svante Eriksen
}

\seealso{
\code{\link{disclapmix}}
\code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}}
\code{\link{summary.disclapmixfit}}
\code{\link{simulate.disclapmixfit}}
\code{\link{haplotype_diversity}}
\code{\link{clusterdist}}
}
\keyword{print}

