% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{print.ranked_contrib_pairs}
\alias{print.ranked_contrib_pairs}
\title{Print ranked contributor pairs}
\usage{
\method{print}{ranked_contrib_pairs}(x, top = 5L,
  hide_non_varying_loci = TRUE, ...)
}
\arguments{
\item{x}{A \code{ranked_contrib_pairs} object.}

\item{top}{The top ranked number of pairs to print/plot. \code{NULL} for
all.}

\item{hide_non_varying_loci}{Whether to hide alleles on loci that do not
vary.}

\item{\dots}{Ignored}
}
\description{
Print ranked contributor pairs
}
