\name{summary.disclapmixfit}
\alias{summary.disclapmixfit}
\title{Summary of a disclapmixfit}
\description{Summary of a \code{\link{disclapmixfit}} object.}
\usage{
  \method{summary}{disclapmixfit}(object, \dots)
}
\arguments{
  \item{object}{a \code{\link{disclapmixfit}} object, usually from a result of a call to \code{disclapmix}.}
  \item{...}{not used}
}
\author{
  Mikkel Meyer Andersen <mikl@math.aau.dk> and Poul Svante Eriksen
}

\seealso{
\code{\link{disclapmix}}
\code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}}
\code{\link{print.disclapmixfit}}
\code{\link{simulate.disclapmixfit}}
%\code{\link{haplotype_diversity}}
\code{\link{clusterdist}}
}
\keyword{print}

