% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterdist.R
\name{clusterdist}
\alias{clusterdist}
\title{Calculate distance between clusters}
\usage{
clusterdist(fit, ...)
}
\arguments{
\item{fit}{A \code{\link{disclapmixfit}} object.}

\item{...}{Not used}
}
\value{
A distance matrix
}
\description{
\code{clusterdist} calculates the distance between each pair of clusters.
The distance measure is based on a symmetric Kullback-Leibler divergence.
}
\seealso{
\code{\link{disclapmix-package}} \code{\link{disclapmix}}
\code{\link{disclapmixfit}} \code{\link{clusterprob}}
\code{\link{predict.disclapmixfit}} \code{\link{print.disclapmixfit}}
\code{\link{summary.disclapmixfit}} \code{\link{simulate.disclapmixfit}}
%\code{\link{haplotype_diversity}}
\code{\link[disclap:disclap-package]{disclap}}
}
\keyword{clusters}
\keyword{distance}
