% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclapmix-package.R
\docType{data}
\name{danes}
\alias{danes}
\title{Y-STR haplotypes}
\format{A data frame with 185 observations on the following 10 loci
(\code{n} is the number of times each haplotype has been observed)
\describe{ \item{DYS19}{} \item{DYS389I}{}
\item{DYS389II}{} \item{DYS390}{} \item{DYS391}{}
\item{DYS392}{} \item{DYS393}{} \item{DYS437}{}
\item{DYS438}{} \item{DYS439}{} \item{n}{} }}
\source{
''Y-chromosome STR haplotypes Danes'' by Hallenberg et al (2005),
http://www.sciencedirect.com/science/article/pii/S0379073805000095
}
\description{
185 Y-STR 10 loci haplotypes
}
\keyword{datasets}
