\name{disclapmix-package}
\alias{disclapmix-package}
\docType{package}
\title{Discrete Laplace mixture inference using the EM algorithm}
\description{\code{disclapmix} makes inference in a mixture of Discrete Laplace distributions using the EM algorithm. A central object is \code{\link{disclapmixfit}}.}
\author{
Mikkel Meyer Andersen and Poul Svante Eriksen

Maintainer: Mikkel Meyer Andersen <mikl@math.aau.dk>
}

\keyword{disclap}
\keyword{disclapmix}
\keyword{disclapmixfit}
\keyword{discrete Laplace}
\keyword{EM algorithm}

\seealso{
\code{\link{disclapmix}}
\code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}}
\code{\link{summary.disclapmixfit}}
\code{\link{simulate.disclapmixfit}}
%\code{\link{haplotype_diversity}}
\code{\link{clusterdist}}
\code{\link[disclap:disclap-package]{disclap}}
}

