% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.disclapmixfit}
\alias{print.disclapmixfit}
\title{Print a disclapmixfit}
\usage{
\method{print}{disclapmixfit}(x, ...)
}
\arguments{
\item{x}{a \code{\link{disclapmixfit}} object, usually from a result of a
call to \code{disclapmix}.}

\item{...}{not used}
}
\description{
Prints a \code{\link{disclapmixfit}} object.
}
\seealso{
\code{\link{disclapmix}} \code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}} \code{\link{summary.disclapmixfit}}
\code{\link{simulate.disclapmixfit}} \code{\link{plot.disclapmixfit}}
%\code{\link{haplotype_diversity}} \code{\link{clusterdist}}
}
\keyword{print}
