% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{simulate.disclapmixfit}
\alias{simulate.disclapmixfit}
\title{Simulate from a disclapmixfit}
\usage{
\method{simulate}{disclapmixfit}(object, nsim = 1L, seed = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{disclapmixfit}} object, usually from a result of
a call to \code{disclapmix}.}

\item{nsim}{number of haplotypes to generate.}

\item{seed}{not used}

\item{...}{not used}
}
\value{
A matrix where the rows correspond to the simulated haplotypes.
}
\description{
Simulate from a \code{\link{disclapmixfit}} object.
}
\seealso{
\code{\link{disclapmix}} \code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}} \code{\link{print.disclapmixfit}}
\code{\link{plot.disclapmixfit}} \code{\link{summary.disclapmixfit}}
%\code{\link{haplotype_diversity}} \code{\link{clusterdist}}
}
\keyword{print}
