% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{summary.disclapmixfit}
\alias{summary.disclapmixfit}
\title{Summary of a disclapmixfit}
\usage{
\method{summary}{disclapmixfit}(object, ...)
}
\arguments{
\item{object}{a \code{\link{disclapmixfit}} object, usually from a result of
a call to \code{disclapmix}.}

\item{...}{not used}
}
\description{
Summary of a \code{\link{disclapmixfit}} object.
}
\seealso{
\code{\link{disclapmix}} \code{\link{disclapmixfit}}
\code{\link{predict.disclapmixfit}} \code{\link{print.disclapmixfit}}
\code{\link{simulate.disclapmixfit}} %\code{\link{haplotype_diversity}}
\code{\link{clusterdist}}
}
\keyword{print}
